+++ drux/messages.py
@@ -17,8 +17,18 @@ ERROR_TARGET_RELEASE_EXCEEDS_MAX = (
 ERROR_INVALID_DIFFUSION = "Diffusivity (D) must be positive."
 ERROR_INVALID_CONCENTRATION = "Initial drug concentration (c0) must be positive."
 ERROR_INVALID_SOLUBILITY = "Solubility (cs) must be positive."
-ERROR_SOLUBILITY_HIGHER_THAN_CONCENTRATION = "Solubility (cs) must be lower or equal to initial concentration (c0)."
+ERROR_SOLUBILITY_HIGHER_THAN_CONCENTRATION = (
+    "Solubility (cs) must be lower or equal to initial concentration (c0)."
+)
 
 # Error messages for zero-order
 ERROR_ZERO_ORDER_RELEASE_RATE = "Release rate (k0) must be non-negative."
-ERROR_ZERO_ORDER_INITIAL_AMOUNT = "Initial amount of drug in the solution (M0) must be non-negative."
+ERROR_ZERO_ORDER_INITIAL_AMOUNT = (
+    "Initial amount of drug in the solution (M0) must be non-negative."
+)
+
+# Error messages for first-order
+ERROR_FIRST_ORDER_RELEASE_RATE = "Release rate (k) must be non-negative."
+ERROR_FIRST_ORDER_INITIAL_AMOUNT = (
+    "Entire releasable amount of drug (M0) must be non-negative."
+)